+++ database/migrations/2025_12_01_123902_create_courses_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('courses')) {
+            return;
+        }
         Schema::create('courses', function (Blueprint $table) {
             $table->id();
             $table->foreignId('university_id')->constrained()->onDelete('cascade');
