+++ database/migrations/2025_12_01_131229_create_packages_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('packages')) {
+            return;
+        }
         Schema::create('packages', function (Blueprint $table) {
             $table->id();
             $table->foreignId('service_module_id')->nullable()->constrained('service_modules')->onDelete('cascade');
